loput osan 4 tehtävät

+++ osa4/blogilista/controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', {username: 1, name: 1, id: 1})
@@ -10,8 +12,13 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
-    const user = (await User.find({}))[0]
-    
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = request.user
+
     const blog = new Blog({
       title: body.title,
       author: body.author,
@@ -32,12 +39,25 @@ blogsRouter.post('/', async (request, response) => {
 })
  
 blogsRouter.delete('/:id', async (request, response) => {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return response.status(401).json({ error: 'invalid token'})
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
-    const userId = await User.find({})[0]._i
+    const userId = (await User.find({}))[0]._id
     
     const blog = {
         title: request.body.title,
